Implemented interceptor code, still failing on return jmp
The example game is Hitman as it allows attaching a debugger and we know
the camera addresses/code addresses so we can finetune the camera system
on that. The code currently intercepts the camera matrix properly but
fails to return back to the right address due to a wrong asm statement.
(see interceptor.asm)