Prey 2017 camera start. Camera works but initial coords read is buggy (race condition)
The problem is that the writes influence each other. Write 4-5 aren't
necessary per se but if e.g. write3 is intercepted before write 1, the
coords start to go to 0. We therefore have to read the initial coords
from somewhere else. E.g. from the camera struct but from a different
offset.