Shader toggler works abit, but hits a major roadblock
It can only work if we can grab something unique and persistent about a shader. This is only the case (to my knowledge) if we can intercept the shader's creation with CreatePixelShader. However the camera tools likely are injected too late and the shader is already created. The shader then can't be identified over multiple game runs, and we can't built a reliable hud toggle.

We can use this machinery to find where calls to PSSetShader originate from, find which shader is part of the hud rendering, and track down its location in memory, then find which code reads from that location and disable it at will when the hud has to be toggled.

Other than that, it's of no use. At least at the moment.