ACO: added critical sections, still failure in optimized release builds
It doesn't make any sense. Debugging in release build in asm, it simply calls all of a sudden a weird location. If it's not at position X then it's at position Y. Could be the DRM in the game, no idea, but it's not something we can fix at our end. Perhaps vs2017 fixes it if it's a bug in vs2015, other than that, it's a mystery why it fails.