SOTTR: Camera done. everything works
For this game, in-lining of functions has been disabled in the C++ optimizations, as it otherwise causes problems in release builds. The cause I don't know but I suspect the code in this file. I've done everything I could to make things work here, but it appears things are still shaky at times. At least with disabling that particular optimization it works
ok.