Everything works now, help shows key bindings too. You can't bind multiple actions to the same key properly tho
This is caused by the sleep() calls to prevent hammering. This sleep should be moved towards the end of the input handling: if an action was handled, it should apply a sleep so no hammering is occurring, otherwise it should not do any sleep(), so it can handle multiple actions per loop.