First commit for NamedPipe/.NET GUI for IGCS (in SOTTR camera).
The plan is to create a .NET based GUI for IGCS, communicating over a named pipe with the injected dll. The .NET based gui will inject the dll, manage the settings and offer controls for the user to control the configuration of the dll, offer feedback from the dll, show notifications and offer hotsampling support (from SRWE).